labs/lab_14_databases/requirements.txt
pytest==7.4.3
rich==13.7.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0 